// typings/api/graphql.tsx
export type URI = any;

/* A string containing HTML code. */
export type HTML = any;

/* An ISO-8601 encoded UTC date string. */
export type DateTime = any;
// ...
export interface Node {
    id: string /* ID of the object. */;
}
/* Represents an object which can take actions on GitHub. Typically a User or Bot. */
export interface Actor {
    avatarUrl: URI /* A URL pointing to the actor&#x27;s public avatar. */;
    login: string /* The username of the actor. */;
    resourcePath: URI /* The HTTP path for this actor. */;
    url: URI /* The HTTP URL for this actor. */;
}
// ...
export interface RepositoryOwner {
    avatarUrl: URI /* A URL pointing to the owner&#x27;s public avatar. */;
    id: string;
    login: string /* The username used to login. */;
    pinnedRepositories: RepositoryConnection /* A list of repositories this user has pinned to their profile */;
    repositories: RepositoryConnection /* A list of repositories that the user owns. */;
    repository?: Repository | null /* Find Repository. */;
    resourcePath: URI /* The HTTP URL for the owner. */;
    url: URI /* The HTTP URL for the owner. */;
}
// ...
export interface UniformResourceLocatable {
    resourcePath: URI /* The HTML path to this resource. */;
    url: URI /* The URL to this resource. */;
}
// ...
export interface Organization extends Node, Actor, ProjectOwner, RepositoryOwner, UniformResourceLocatable {
    avatarUrl: URI /* A URL pointing to the organization&#x27;s public avatar. */;
    databaseId?: number | null /* Identifies the primary key from the database. */;
    description?: string | null /* The organization&#x27;s public profile description. */;
    email?: string | null /* The organization&#x27;s public email. */;
    id: string;
    location?: string | null /* The organization&#x27;s public profile location. */;
    login: string /* The organization&#x27;s login name. */;
    members: UserConnection /* A list of users who are members of this organization. */;
    name?: string | null /* The organization&#x27;s public profile name. */;
    newTeamResourcePath: URI /* The HTTP path creating a new team */;
    newTeamUrl: URI /* The HTTP URL creating a new team */;
    organizationBillingEmail?: string | null /* The billing email for the organization. */;
    pinnedRepositories: RepositoryConnection /* A list of repositories this user has pinned to their profile */;
    project?: Project | null /* Find project by number. */;
    projects: ProjectConnection /* A list of projects under the owner. */;
    projectsResourcePath: URI /* The HTTP path listing organization&#x27;s projects */;
    projectsUrl: URI /* The HTTP URL listing organization&#x27;s projects */;
    repositories: RepositoryConnection /* A list of repositories that the user owns. */;
    repository?: Repository | null /* Find Repository. */;
    resourcePath: URI /* The HTTP path for this user */;
    samlIdentityProvider?: OrganizationIdentityProvider | null /* The Organization&#x27;s SAML Identity Providers */;
    team?: Team | null /* Find an organization&#x27;s team by its slug. */;
    teams: TeamConnection /* A list of teams in this organization. */;
    teamsResourcePath: URI /* The HTTP path listing organization&#x27;s teams */;
    teamsUrl: URI /* The HTTP URL listing organization&#x27;s teams */;
    url: URI /* The HTTP URL for this user */;
    viewerCanAdminister: boolean /* Organization is adminable by the viewer. */;
    viewerCanCreateProjects: boolean /* Can the current viewer create new projects on this owner. */;
    viewerCanCreateRepositories: boolean /* Viewer can create repositories on this organization */;
    viewerCanCreateTeams: boolean /* Viewer can create teams on this organization. */;
    viewerIsAMember: boolean /* Viewer is an active member of this organization. */;
    websiteUrl?: URI | null /* The organization&#x27;s public profile URL. */;
}
// ...
export interface User extends Node, Actor, RepositoryOwner, UniformResourceLocatable {
    avatarUrl: URI /* A URL pointing to the user&#x27;s public avatar. */;
    bio?: string | null /* The user&#x27;s public profile bio. */;
    bioHTML: HTML /* The user&#x27;s public profile bio as HTML. */;
    commitComments: CommitCommentConnection /* A list of commit comments made by this user. */;
    company?: string | null /* The user&#x27;s public profile company. */;
    companyHTML: HTML /* The user&#x27;s public profile company as HTML. */;
    contributedRepositories: RepositoryConnection /* A list of repositories that the user recently contributed to. */;
    createdAt: DateTime /* Identifies the date and time when the object was created. */;
    databaseId?: number | null /* Identifies the primary key from the database. */;
    email: string /* The user&#x27;s publicly visible profile email. */;
    followers: FollowerConnection /* A list of users the given user is followed by. */;
    following: FollowingConnection /* A list of users the given user is following. */;
    gist?: Gist | null /* Find gist by repo name. */;
    gistComments: GistCommentConnection /* A list of gist comments made by this user. */;
    gists: GistConnection /* A list of the Gists the user has created. */;
    id: string;
    isBountyHunter: boolean /* Whether or not this user is a participant in the GitHub Security Bug Bounty. */;
    isCampusExpert: boolean /* Whether or not this user is a participant in the GitHub Campus Experts Program. */;
    isDeveloperProgramMember: boolean /* Whether or not this user is a GitHub Developer Program member. */;
    isEmployee: boolean /* Whether or not this user is a GitHub employee. */;
    isHireable: boolean /* Whether or not the user has marked themselves as for hire. */;
    isSiteAdmin: boolean /* Whether or not this user is a site administrator. */;
    isViewer: boolean /* Whether or not this user is the viewing user. */;
    issueComments: IssueCommentConnection /* A list of issue comments made by this user. */;
    issues: IssueConnection /* A list of issues assocated with this user. */;
    location?: string | null /* The user&#x27;s public profile location. */;
    login: string /* The username used to login. */;
    name?: string | null /* The user&#x27;s public profile name. */;
    organization?: Organization | null /* Find an organization by its login that the user belongs to. */;
    organizations: OrganizationConnection /* A list of organizations the user belongs to. */;
    pinnedRepositories: RepositoryConnection /* A list of repositories this user has pinned to their profile */;
    publicKeys: PublicKeyConnection /* A list of public keys associated with this user. */;
    pullRequests: PullRequestConnection /* A list of pull requests assocated with this user. */;
    repositories: RepositoryConnection /* A list of repositories that the user owns. */;
    repositoriesContributedTo: RepositoryConnection /* A list of repositories that the user recently contributed to. */;
    repository?: Repository | null /* Find Repository. */;
    resourcePath: URI /* The HTTP path for this user */;
    starredRepositories: StarredRepositoryConnection /* Repositories the user has starred. */;
    updatedAt: DateTime /* Identifies the date and time when the object was last updated. */;
    url: URI /* The HTTP URL for this user */;
    viewerCanFollow: boolean /* Whether or not the viewer is able to follow the user. */;
    viewerIsFollowing: boolean /* Whether or not this user is followed by the viewer. */;
    watching: RepositoryConnection /* A list of repositories the given user is watching. */;
    websiteUrl?: URI | null /* A URL pointing to the user&#x27;s public website/blog. */;
}
/* A list of repositories owned by the subject. */
export interface RepositoryConnection {
    edges?: RepositoryEdge[] | null /* A list of edges. */;
    nodes?: Repository[] | null /* A list of nodes. */;
    pageInfo: PageInfo /* Information to aid in pagination. */;
    totalCount: number /* Identifies the total count of items in the connection. */;
    totalDiskUsage: number /* The total size in kilobytes of all repositories in the connection. */;
}
// ...
export interface CommitCommentConnection {
    edges?: CommitCommentEdge[] | null /* A list of edges. */;
    nodes?: CommitComment[] | null /* A list of nodes. */;
    pageInfo: PageInfo /* Information to aid in pagination. */;
    totalCount: number /* Identifies the total count of items in the connection. */;
}
/* An edge in a connection. */
export interface CommitCommentEdge {
    cursor: string /* A cursor for use in pagination. */;
    node?: CommitComment | null /* The item at the end of the edge. */;
}
// ...
export interface PullRequestConnection {
    edges?: PullRequestEdge[] | null /* A list of edges. */;
    nodes?: PullRequest[] | null /* A list of nodes. */;
    pageInfo: PageInfo /* Information to aid in pagination. */;
    totalCount: number /* Identifies the total count of items in the connection. */;
}
// ...
export interface IssueConnection {
    edges?: IssueEdge[] | null /* A list of edges. */;
    nodes?: Issue[] | null /* A list of nodes. */;
    pageInfo: PageInfo /* Information to aid in pagination. */;
    totalCount: number /* Identifies the total count of items in the connection. */;
}
/* An edge in a connection. */
export interface IssueEdge {
    cursor: string /* A cursor for use in pagination. */;
    node?: Issue | null /* The item at the end of the edge. */;
}
/* The connection type for IssueComment. */
export interface IssueCommentConnection {
    edges?: IssueCommentEdge[] | null /* A list of edges. */;
    nodes?: IssueComment[] | null /* A list of nodes. */;
    pageInfo: PageInfo /* Information to aid in pagination. */;
    totalCount: number /* Identifies the total count of items in the connection. */;
}
// ...
export interface FollowerConnection {
    edges?: UserEdge[] | null /* A list of edges. */;
    nodes?: User[] | null /* A list of nodes. */;
    pageInfo: PageInfo /* Information to aid in pagination. */;
    totalCount: number /* Identifies the total count of items in the connection. */;
}
/* The connection type for User. */
export interface FollowingConnection {
    edges?: UserEdge[] | null /* A list of edges. */;
    nodes?: User[] | null /* A list of nodes. */;
    pageInfo: PageInfo /* Information to aid in pagination. */;
    totalCount: number /* Identifies the total count of items in the connection. */;
}
/* A Gist. */
export interface Gist extends Node, Starrable {
    comments: GistCommentConnection /* A list of comments associated with the gist */;
    createdAt: DateTime /* Identifies the date and time when the object was created. */;
    description?: string | null /* The gist description. */;
    id: string;
    isPublic: boolean /* Whether the gist is public or not. */;
    name: string /* The gist name. */;
    owner?: RepositoryOwner | null /* The gist owner. */;
    pushedAt?: DateTime | null /* Identifies when the gist was last pushed to. */;
    stargazers: StargazerConnection /* A list of users who have starred this starrable. */;
    updatedAt: DateTime /* Identifies the date and time when the object was last updated. */;
    viewerHasStarred: boolean /* Returns a boolean indicating whether the viewing user has starred this starrable. */;
}
/* The connection type for GistComment. */
export interface GistCommentConnection {
    edges?: GistCommentEdge[] | null /* A list of edges. */;
    nodes?: GistComment[] | null /* A list of nodes. */;
    pageInfo: PageInfo /* Information to aid in pagination. */;
    totalCount: number /* Identifies the total count of items in the connection. */;
}
// ...
export interface GistConnection {
    edges?: GistEdge[] | null /* A list of edges. */;
    nodes?: Gist[] | null /* A list of nodes. */;
    pageInfo: PageInfo /* Information to aid in pagination. */;
    totalCount: number /* Identifies the total count of items in the connection. */;
}
/* An edge in a connection. */
export interface GistEdge {
    cursor: string /* A cursor for use in pagination. */;
    node?: Gist | null /* The item at the end of the edge. */;
}
/* The connection type for Organization. */
export interface OrganizationConnection {
    edges?: OrganizationEdge[] | null /* A list of edges. */;
    nodes?: Organization[] | null /* A list of nodes. */;
    pageInfo: PageInfo /* Information to aid in pagination. */;
    totalCount: number /* Identifies the total count of items in the connection. */;
}
/* An edge in a connection. */
export interface OrganizationEdge {
    cursor: string /* A cursor for use in pagination. */;
    node?: Organization | null /* The item at the end of the edge. */;
}
/* The connection type for PublicKey. */
export interface PublicKeyConnection {
    edges?: PublicKeyEdge[] | null /* A list of edges. */;
    nodes?: PublicKey[] | null /* A list of nodes. */;
    pageInfo: PageInfo /* Information to aid in pagination. */;
    totalCount: number /* Identifies the total count of items in the connection. */;
}